+++ gmd/Cui/FilterDlg.cs
@@ -11,7 +11,7 @@ interface IFilterDlg
 
 class FilterDlg : IFilterDlg
 {
-    const int MaxResults = int.MaxValue;
+    const int MaxResults = 5000;
     readonly IServer server;
     readonly IBranchColorService branchColorService;
 
@@ -50,8 +50,8 @@ class FilterDlg : IFilterDlg
 
         dlg.AddLabel(0, 0, Text.BrightMagenta("Gmd"));
 
-        dlg.AddLabel(Application.Driver.Cols - 41, 0, Text.BrightMagenta("Search"));
-        filterField = dlg.AddTextField(Application.Driver.Cols - 34, 0, 30);
+        dlg.AddLabel(Application.Driver.Cols - 41, 0, Text.BrightMagenta("Search:"));
+        filterField = dlg.AddInputField(Application.Driver.Cols - 33, 0, 29);
         filterField.KeyUp += (k) => OnFilterFieldKeyUp(k);    // Update results and select commit on keys
 
         statusLabel = dlg.AddLabel(6, 0);