+++ gmd/Cui/RepoView.cs
@@ -61,7 +61,8 @@ class RepoView : IRepoView
     bool isShowFilter;
     bool isRegistered = false;
     string hooverBranchName = "";
-    int hooverIndex = -1;
+    int hooverRowIndex = -1;
+    private int hooverColumnIndex;
     int hooverCurrentCommitIndex = -1;
 
     internal RepoView(
@@ -200,6 +201,7 @@ class RepoView : IRepoView
         commitsView.IsFocus = true;
         commitsView.SetFocus();
         commitsView.SetNeedsDisplay();
+        Application.Driver.SetCursorVisibility(CursorVisibility.Invisible);
 
         if (commit != null)
         {   // User selected a commit, show it
@@ -281,6 +283,7 @@ class RepoView : IRepoView
         commitsView.RegisterKeyHandler(Key.CursorRight, () => OnCursorRight());
         commitsView.RegisterKeyHandler(Key.CursorUp, () => OnCursorUp());
         commitsView.RegisterKeyHandler(Key.CursorDown, () => OnCursorDown());
+        commitsView.RegisterKeyHandler(Key.CursorRight | Key.ShiftMask, OnKeyShiftCursorRight);
 
         commitsView.RegisterKeyHandler(Key.r, () => RefreshAndFetch());
         commitsView.RegisterKeyHandler(Key.F5, () => RefreshAndFetch());
@@ -288,7 +291,7 @@ class RepoView : IRepoView
         commitsView.RegisterKeyHandler(Key.a, () => Cmd.Commit(true));
         commitsView.RegisterKeyHandler(Key.t, () => Cmd.AddTag());
         commitsView.RegisterKeyHandler(Key.b, () => Cmd.CreateBranch());
-        commitsView.RegisterKeyHandler(Key.d, () => Cmd.ShowCurrentRowDiff());
+        commitsView.RegisterKeyHandler(Key.d, OnKeyD);
         commitsView.RegisterKeyHandler(Key.D | Key.CtrlMask, () => Cmd.ShowCurrentRowDiff());
         commitsView.RegisterKeyHandler(Key.p, () => Cmd.PushCurrentBranch());
         commitsView.RegisterKeyHandler(Key.P, () => Cmd.PushAllBranches());
@@ -299,6 +302,7 @@ class RepoView : IRepoView
         commitsView.RegisterKeyHandler((Key)63, () => Cmd.ShowHelp()); // '?' key
         commitsView.RegisterKeyHandler(Key.f, () => OnKeyF());
         commitsView.RegisterKeyHandler(Key.D0, () => charDlg.Show());
+        commitsView.RegisterKeyHandler(Key.D5, () => Cmd.SetBranchManuallyAsync());
 
         commitsView.RegisterKeyHandler(Key.y, () => Cmd.ShowBranch(repo.GetCurrentBranch().Name, false));
         commitsView.RegisterKeyHandler(Key.s, () => OnKeyS());
@@ -323,6 +327,24 @@ class RepoView : IRepoView
         applicationBarView.ItemClicked += OnApplicationClick;
     }
 
+    void OnKeyShiftCursorRight()
+    {
+        var x = hooverBranchName == "" ? repo.Graph.Width + 2 : hooverColumnIndex;
+        var y = hooverRowIndex == -1 ? repo.CurrentIndex : hooverRowIndex;
+        menuService.ShowOpenBranchesMenu(x + 2, y + 1);
+    }
+
+    void OnKeyD()
+    {
+        if (hooverBranchName != "")
+        {
+            menuService.ShowDiffBranchToMenu(hooverColumnIndex + 2, hooverRowIndex + 1, hooverBranchName);
+            return;
+        }
+
+        Cmd.ShowCurrentRowDiff();
+    }
+
     void OnKeyF()
     {
         ClearHoover();
@@ -337,10 +359,10 @@ class RepoView : IRepoView
                 Cmd.UpdateRelease();
                 break;
             case ApplicationBarItem.Gmd:
-                menuService.ShowRepoMenu(x - 5, y + 2);
+                menuService.ShowRepoMenu(x - 5, y);
                 break;
             case ApplicationBarItem.Repo:
-                menuService.ShowOpenRepoMenu(x - 5, y + 2);
+                menuService.ShowOpenRepoMenu(x - 5, y);
                 break;
             case ApplicationBarItem.CurrentBranch:
                 Cmd.ShowBranch(repo.GetCurrentBranch().Name, false);
@@ -355,7 +377,7 @@ class RepoView : IRepoView
                 Cmd.PushAllBranches();
                 break;
             case ApplicationBarItem.BranchName:
-                menuService.ShowOpenBranchesMenu(x - 5, y + 2);
+                menuService.ShowOpenBranchesMenu(x - 5, y);
                 break;
             case ApplicationBarItem.Search:
                 Cmd.Filter();
@@ -690,10 +712,10 @@ class RepoView : IRepoView
             {
                 ClearHoover();
             }
-            hooverIndex = currentIndex;
+            hooverRowIndex = currentIndex;
         }
 
-        return (repoWriter.ToPage(repo, firstIndex, count, currentIndex, hooverBranchName, hooverIndex, width), repo.Commits.Count);
+        return (repoWriter.ToPage(repo, firstIndex, count, currentIndex, hooverBranchName, hooverRowIndex, width), repo.Commits.Count);
     }
 
 
@@ -721,7 +743,8 @@ class RepoView : IRepoView
         {
             var index = y + commitsView.FirstIndex;
             hooverBranchName = "";
-            hooverIndex = index;
+            hooverRowIndex = index;
+            hooverColumnIndex = repo.Graph.Width + 2;
             hooverCurrentCommitIndex = repo.CurrentIndex;
             commitsView.SetNeedsDisplay();
             return;
@@ -731,10 +754,11 @@ class RepoView : IRepoView
 
     void SetHooverBranch(GraphBranch branch, int index)
     {
-        if (hooverBranchName != branch.B.PrimaryName || index != hooverIndex)
+        if (hooverBranchName != branch.B.PrimaryName || index != hooverRowIndex || branch.X * 2 != hooverColumnIndex)
         {
             hooverBranchName = branch.B.PrimaryName;
-            hooverIndex = index;
+            hooverRowIndex = index;
+            hooverColumnIndex = branch.X * 2;
             hooverCurrentCommitIndex = repo.CurrentIndex;
             applicationBarView.SetBranch(branch);
             commitsView.SetNeedsDisplay();
@@ -743,10 +767,11 @@ class RepoView : IRepoView
 
     void ClearHoover()
     {
-        if (hooverBranchName != "" || hooverIndex != -1 || hooverCurrentCommitIndex != -1)
+        if (hooverBranchName != "" || hooverRowIndex != -1 || hooverCurrentCommitIndex != -1 || hooverColumnIndex != -1)
         {
             hooverBranchName = "";
-            hooverIndex = -1;
+            hooverRowIndex = -1;
+            hooverColumnIndex = -1;
             hooverCurrentCommitIndex = -1;
             commitsView.SetNeedsDisplay();
         }
@@ -870,7 +895,6 @@ class RepoView : IRepoView
 
         var names = repo.Branches.Select(b => b.PrimaryBaseName).Distinct().Take(30).ToList();
         repoState.Set(serverRepo.Path, s => s.Branches = names);
-
     }
 
 
@@ -915,6 +939,8 @@ class RepoView : IRepoView
 
     void OnCurrentIndexChange()
     {
+        if (repo.CurrentIndex < 0) return;
+
         var commit = repo.RowCommit;
         var branch = repo.Graph.BranchByName(commit.BranchName);
         applicationBarView.SetBranch(branch);